refactored

+++ backend/repositories/authRepository.js
@@ -11,46 +11,68 @@ class AuthRepository extends Repository {
     const credential = await db.Credential.findOne({
       where: {
         email,
+        role,
       },
-      include: [{ model: db.User, required: true, where: { role } }],
+      include: [{ model: db.User, required: true }],
     });
     return credential;
   };
 
   getUserByNameType = async (username, role) => {
     const credential = await db.Credential.findOne({
-      include: [{ model: db.User, required: true, where: { username, role } }],
+      where: {
+        role,
+      },
+      include: [{ model: db.User, required: true, where: { username } }],
     });
     return credential;
   };
 
   deleteAccount = async (id) => {
-    const deletedUser = await db.User.destroy({
-      where: {
-        id,
-      },
-      returning: true, // This option returns the deleted user
-    });
-
-    if (!deletedUser) {
-      return null; // Handle the case where the user does not exist or no rows were deleted
+    const user = await db.User.findByPk(id);
+    if (user) {
+      await db.User.destroy({
+        where: {
+          id,
+        },
+      });
+      return user;
     }
-
-    return deletedUser;
+    return null;
   };
 
   signup = async (data) => {
-    const user = await db.User.create({
-      username: data.username,
-      fullname: data.fullname,
-      role: data.type,
-    });
-    await db.Credential.create({
-      userId: user.id,
-      email: data.email,
-      hashpass: data.hashPass,
-    });
-    return user;
+    console.log(data);
+    let transaction;
+    try {
+      transaction = await db.sequelize.transaction();
+      const user = await db.User.create(
+        {
+          username: data.username,
+          fullname: data.fullname,
+        },
+        { transaction }
+      );
+      await db.Credential.create(
+        {
+          userId: user.id,
+          email: data.email,
+          hashpass: data.hashPass,
+          role: data.type,
+        },
+        { transaction }
+      );
+      await transaction.commit();
+      return user;
+    } catch (err) {
+      // If an error occurs, rollback the transaction
+      if (transaction) await transaction.rollback();
+      if (err.name === "SequelizeUniqueConstraintError") {
+        return null;
+      }
+      // Handle the error or rethrow it
+      throw err;
+    }
   };
 }
